stas: refactoring code to allow for module override
