+++ staslib/stas.py
@@ -6,14 +6,19 @@
 #
 # Authors: Martin Belanger <Martin.Belanger@dell.com>
 #
-'''Library for staf/stac'''
+'''Library for staf/stac. You will find here common code for stafd and stacd
+including the Abstract Base Classes (ABC) for Controllers and Services'''
 
 import os
 import sys
-import ipaddress
+import abc
+import signal
 import logging
-
-from staslib import conf, defs, trid
+import ipaddress
+import systemd.daemon
+import dasbus.connection
+from gi.repository import Gio, GLib
+from staslib import conf, defs, gutil, log, trid
 
 
 # ******************************************************************************
@@ -108,3 +113,379 @@ def remove_invalid_addresses(controllers: list):
             logging.warning('Invalid transport %s', transport)
 
     return valid_controllers
+
+
+# ******************************************************************************
+class ControllerABC(abc.ABC):  # pylint: disable=too-many-instance-attributes
+    '''@brief Base class used to manage the connection to a controller.'''
+
+    CONNECT_RETRY_PERIOD_SEC = 60
+    FAST_CONNECT_RETRY_PERIOD_SEC = 3
+
+    def __init__(self, root, host, tid: trid.TID, discovery_ctrl=False):
+        self._root              = root
+        self._host              = host
+        self._tid               = tid
+        self._cancellable       = Gio.Cancellable()
+        self._connect_attempts  = 0
+        self._retry_connect_tmr = gutil.GTimer(self.CONNECT_RETRY_PERIOD_SEC, self._on_try_to_connect)
+        self._discovery_ctrl    = discovery_ctrl
+        self._try_to_connect_deferred = gutil.Deferred(self._try_to_connect)
+        self._try_to_connect_deferred.schedule()
+
+    def _release_resources(self):
+        # Remove pending deferred from main loop
+        if self._try_to_connect_deferred:
+            self._try_to_connect_deferred.cancel()
+
+        if self._retry_connect_tmr is not None:
+            self._retry_connect_tmr.kill()
+
+        if self._cancellable and not self._cancellable.is_cancelled():
+            self._cancellable.cancel()
+
+        self._tid = None
+        self._cancellable = None
+        self._retry_connect_tmr = None
+        self._try_to_connect_deferred = None
+
+    @property
+    def id(self) -> str:
+        '''@brief Return the Transport ID as a printable string'''
+        return str(self.tid)
+
+    @property
+    def tid(self):
+        '''@brief Return the Transport ID object'''
+        return self._tid
+
+    def controller_id_dict(self) -> dict:
+        '''@brief return the controller ID as a dict.'''
+        return self.tid.as_dict()
+
+    def details(self) -> dict:
+        '''@brief return detailed debug info about this controller'''
+        details = self.controller_id_dict()
+        details['connect attempts'] = str(self._connect_attempts)
+        details['retry connect timer'] = str(self._retry_connect_tmr)
+        return details
+
+    def info(self) -> dict:
+        '''@brief Get the controller info for this object'''
+        return self.details()
+
+    def cancel(self):
+        '''@brief Used to cancel pending operations.'''
+        if self._cancellable and not self._cancellable.is_cancelled():
+            logging.debug('ControllerABC.cancel()             - %s', self.id)
+            self._cancellable.cancel()
+
+    def kill(self):
+        '''@brief Used to release all resources associated with this object.'''
+        logging.debug('ControllerABC.kill()               - %s', self.id)
+        self._release_resources()
+
+    def _alive(self):
+        '''There may be race condition where a queued event gets processed
+        after the object is no longer configured (i.e. alive). This method
+        can be used by callback functions to make sure the object is still
+        alive before processing further.
+        '''
+        return self._cancellable and not self._cancellable.is_cancelled()
+
+    def _on_try_to_connect(self):
+        self._try_to_connect_deferred.schedule()
+        return GLib.SOURCE_REMOVE
+
+    def _try_to_connect(self):
+        # This is a deferred function call. Make sure
+        # the source of the deferred is still good.
+        source = GLib.main_current_source()
+        if source and source.is_destroyed():
+            return
+
+        self._connect_attempts += 1
+
+        self._do_connect()
+
+    @abc.abstractmethod
+    def _do_connect(self):
+        raise NotImplementedError()
+
+    @abc.abstractmethod
+    def _on_aen(self, aen: int):
+        raise NotImplementedError()
+
+    @abc.abstractmethod
+    def _on_nvme_event(self, nvme_event):
+        raise NotImplementedError()
+
+    @abc.abstractmethod
+    def _on_ctrl_removed(self, obj):
+        raise NotImplementedError()
+
+    @abc.abstractmethod
+    def _find_existing_connection(self):
+        raise NotImplementedError()
+
+    @abc.abstractmethod
+    def disconnect(self, disconnected_cb, keep_connection):
+        '''@brief Issue an asynchronous disconnect command to a Controller.
+        Once the async command has completed, the callback 'disconnected_cb'
+        will be invoked. If a controller is already disconnected, then the
+        callback will be added to the main loop's next idle slot to be executed
+        ASAP.
+        '''
+        raise NotImplementedError()
+
+
+# ******************************************************************************
+class ServiceABC(abc.ABC):  # pylint: disable=too-many-instance-attributes
+    '''@brief Base class used to manage a STorage Appliance Service'''
+
+    CONF_STABILITY_SOAK_TIME_SEC = 1.5
+
+    def __init__(self, args, reload_hdlr):
+
+        service_conf = conf.SvcConf()
+        service_conf.set_conf_file(args.conf_file)  # reload configuration
+        self._tron = args.tron or service_conf.tron
+        log.set_level_from_tron(self._tron)
+
+        self._lkc_file     = os.path.join(os.environ.get('RUNTIME_DIRECTORY', os.path.join('/run', defs.PROG_NAME)), 'last-known-config.pickle')
+        self._loop         = GLib.MainLoop()
+        self._cancellable  = Gio.Cancellable()
+        self._resolver     = gutil.NameResolver()
+        self._controllers  = self._load_last_known_config()
+        self._dbus_iface   = None
+        self._cfg_soak_tmr = gutil.GTimer(self.CONF_STABILITY_SOAK_TIME_SEC, self._on_config_ctrls)
+        self._sysbus       = dasbus.connection.SystemMessageBus()
+
+        GLib.unix_signal_add(GLib.PRIORITY_HIGH, signal.SIGINT, self._stop_hdlr)  # CTRL-C
+        GLib.unix_signal_add(GLib.PRIORITY_HIGH, signal.SIGTERM, self._stop_hdlr)  # systemctl stop stafd
+        GLib.unix_signal_add(GLib.PRIORITY_HIGH, signal.SIGHUP, reload_hdlr)  # systemctl reload stafd
+
+        nvme_options = conf.NvmeOptions()
+        if not nvme_options.host_iface_supp or not nvme_options.discovery_supp:
+            logging.warning(
+                'Kernel does not appear to support all the options needed to run this program. Consider updating to a later kernel version.'
+            )
+
+        # We don't want to apply configuration changes to nvme-cli right away.
+        # Often, multiple changes will occur in a short amount of time (sub-second).
+        # We want to wait until there are no more changes before applying them
+        # to the system. The following timer acts as a "soak period". Changes
+        # will be applied by calling self._on_config_ctrls() at the end of
+        # the soak period.
+        self._cfg_soak_tmr.start()
+
+    def _release_resources(self):
+        logging.debug('ServiceABC._release_resources()')
+
+        if self._cancellable and not self._cancellable.is_cancelled():
+            self._cancellable.cancel()
+
+        if self._cfg_soak_tmr is not None:
+            self._cfg_soak_tmr.kill()
+
+        self._controllers.clear()
+
+        if self._sysbus:
+            self._sysbus.disconnect()
+
+        self._cfg_soak_tmr = None
+        self._cancellable = None
+        self._resolver = None
+        self._lkc_file = None
+        self._sysbus = None
+
+    def _config_dbus(self, iface_obj, bus_name: str, obj_name: str):
+        self._dbus_iface = iface_obj
+        self._sysbus.publish_object(obj_name, iface_obj)
+        self._sysbus.register_service(bus_name)
+
+    @property
+    def tron(self):
+        '''@brief Get Trace ON property'''
+        return self._tron
+
+    @tron.setter
+    def tron(self, value):
+        '''@brief Set Trace ON property'''
+        self._tron = value
+        log.set_level_from_tron(self._tron)
+
+    def run(self):
+        '''@brief Start the main loop execution'''
+        try:
+            self._loop.run()
+        except Exception as ex:  # pylint: disable=broad-except
+            logging.critical('exception: %s', ex)
+
+        self._loop = None
+
+    def info(self) -> dict:
+        '''@brief Get the status info for this object (used for debug)'''
+        nvme_options = conf.NvmeOptions()
+        return {
+            'last known config file': self._lkc_file,
+            'config soak timer': str(self._cfg_soak_tmr),
+            'kernel support': {
+                'TP8013': nvme_options.discovery_supp,
+                'host_iface': nvme_options.host_iface_supp,
+            },
+            'system config': conf.SysConf().as_dict(),
+        }
+
+    def get_controllers(self) -> dict:
+        '''@brief return the list of controller objects'''
+        return self._controllers.values()
+
+    def get_controller(
+        self, transport: str, traddr: str, trsvcid: str, host_traddr: str, host_iface: str, subsysnqn: str
+    ):  # pylint: disable=too-many-arguments
+        '''@brief get the specified controller object from the list of controllers'''
+        cid = {
+            'transport': transport,
+            'traddr': traddr,
+            'trsvcid': trsvcid,
+            'host-traddr': host_traddr,
+            'host-iface': host_iface,
+            'subsysnqn': subsysnqn,
+        }
+        return self._controllers.get(trid.TID(cid))
+
+    def _remove_ctrl_from_dict(self, controller):
+        tid_to_pop = controller.tid
+        if not tid_to_pop:
+            # Being paranoid. This should not happen, but let's say the
+            # controller object has been purged, but it is somehow still
+            # listed in self._controllers.
+            for tid, _controller in self._controllers.items():
+                if _controller is controller:
+                    tid_to_pop = tid
+                    break
+
+        if tid_to_pop:
+            logging.debug('ServiceABC._remove_ctrl_from_dict()- %s | %s', tid_to_pop, controller.device)
+            self._controllers.pop(tid_to_pop, None)
+        else:
+            logging.debug('ServiceABC._remove_ctrl_from_dict()- already removed')
+
+    def remove_controller(self, controller, success):  # pylint: disable=unused-argument
+        '''@brief remove the specified controller object from the list of controllers
+        @param controller: the controller object
+        @param success: whether the disconnect was successful'''
+        logging.debug('ServiceABC.remove_controller()')
+        if isinstance(controller, ControllerABC):
+            self._remove_ctrl_from_dict(controller)
+
+            controller.kill()
+
+        if self._cfg_soak_tmr:
+            self._cfg_soak_tmr.start()
+
+    def _cancel(self):
+        logging.debug('ServiceABC._cancel()')
+        if not self._cancellable.is_cancelled():
+            self._cancellable.cancel()
+
+        for controller in self._controllers.values():
+            controller.cancel()
+
+    def _stop_hdlr(self):
+        logging.debug('ServiceABC._stop_hdlr()')
+        systemd.daemon.notify('STOPPING=1')
+
+        self._cancel()  # Cancel pending operations
+
+        self._dump_last_known_config(self._controllers)
+
+        if len(self._controllers) == 0:
+            GLib.idle_add(self._exit)
+        else:
+            # Tell all controller objects to disconnect
+            keep_connections = self._keep_connections_on_exit()
+            controllers = self._controllers.values()
+            logging.debug(
+                'ServiceABC._stop_hdlr()            - Controller count = %s, keep_connections = %s',
+                len(controllers), keep_connections
+            )
+            for controller in controllers:
+                controller.disconnect(self._on_final_disconnect, keep_connections)
+
+        return GLib.SOURCE_REMOVE
+
+    def _on_final_disconnect(self, controller, success):
+        '''Callback invoked after a controller is disconnected.
+        THIS IS USED DURING PROCESS SHUTDOWN TO WAIT FOR ALL CONTROLLERS TO BE
+        DISCONNECTED BEFORE EXITING THE PROGRAM. ONLY CALL ON SHUTDOWN!
+        @param controller: the controller object
+        @param success: whether the disconnect operation was successful
+        '''
+        logging.debug('ServiceABC._on_final_disconnect()  - %s | %s disconnect %s',
+                      controller.id, controller.device, 'succeeded' if success else 'failed')
+        self._remove_ctrl_from_dict(controller)
+
+        controller.kill()
+
+        # When all controllers have disconnected, we can finish the clean up
+        if len(self._controllers) == 0:
+            # Defer exit to the next main loop's idle period.
+            GLib.idle_add(self._exit)
+
+    def _exit(self):
+        logging.debug('ServiceABC._exit()')
+        self._release_resources()
+        self._loop.quit()
+
+    def _on_config_ctrls(self, *_user_data):
+        self._config_ctrls()
+        return GLib.SOURCE_REMOVE
+
+    def _config_ctrls(self):
+        '''@brief Start controllers configuration.'''
+        # The configuration file may contain controllers and/or blacklist
+        # elements with traddr specified as hostname instead of IP address.
+        # Because of this, we need to remove those blacklisted elements before
+        # running name resolution. And we will need to remove blacklisted
+        # elements after name resolution is complete (i.e. in the calback
+        # function _config_ctrls_finish)
+        logging.debug('ServiceABC._config_ctrls()')
+        configured_controllers = remove_blacklisted(conf.SvcConf().get_controllers())
+        self._resolver.resolve_ctrl_async(self._cancellable, configured_controllers, self._config_ctrls_finish)
+
+    @abc.abstractmethod
+    def _keep_connections_on_exit(self):
+        '''@brief Determine whether connections should remain when the
+        process exits.
+
+        NOTE) This is the base class method used to define the interface.
+        It must be overloaded by a child class.
+        '''
+        raise NotImplementedError()
+
+    @abc.abstractmethod
+    def _config_ctrls_finish(self, configured_ctrl_list):
+        '''@brief Finish controllers configuration after hostnames (if any)
+        have been resolved.
+
+        Configuring controllers must be done asynchronously in 2 steps.
+        In the first step, host names get resolved to find their IP addresses.
+        Name resolution can take a while, especially when an external name
+        resolution server is used. Once that step completed, the callback
+        method _config_ctrls_finish() (i.e. this method), gets invoked to
+        complete the controller configuration.
+
+        NOTE) This is the base class method used to define the interface.
+        It must be overloaded by a child class.
+        '''
+        raise NotImplementedError()
+
+    @abc.abstractmethod
+    def _load_last_known_config(self):
+        raise NotImplementedError()
+
+    @abc.abstractmethod
+    def _dump_last_known_config(self, controllers):
+        raise NotImplementedError()
